Refactor trail_info.dart, comment each sub-page with a brief summary
