+++ lib/page/home_page/map_page/trail_cursor_marker_layer.dart
@@ -3,7 +3,7 @@ import 'package:flutter_map/flutter_map.dart';
 import 'package:forest_park_reports/provider/map_cursor_provider.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-
+/// A cursor indicator displayed over a trail when a user drags over the elevation graph
 class TrailCursorMarkerLayer extends ConsumerWidget {
   const TrailCursorMarkerLayer({super.key});
