+++ lib/page/home_page/panel_page/hazard_image.dart
@@ -4,6 +4,7 @@ import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 import 'package:forest_park_reports/provider/hazard_provider.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
+/// Displays an image of a hazard; used in hazard_info, trail_info, and panel_page
 class HazardImage extends ConsumerWidget {
   final String uuid;
   const HazardImage(this.uuid, {super.key});
