+++ lib/page/home_page/panel_page/platform_pill.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
+
+/// A pill that indicates draggability of the main panel
+class PlatformPill extends StatelessWidget {
+  const PlatformPill({super.key});
+  @override
+  Widget build(BuildContext context) {
+    final isIos = isCupertino(context);
+    final theme = Theme.of(context);
+    return Align(
+      alignment: Alignment.topCenter,
+      child: Container(
+        margin: EdgeInsets.symmetric(
+            vertical: isIos ? 5 : 10
+        ),
+        width: isIos ? 35 : 26,
+        height: 5,
+        decoration: BoxDecoration(
+            color: isIos
+                ? CupertinoDynamicColor.resolve(CupertinoColors.systemGrey2, context)
+                : theme.colorScheme.onSurface,
+            borderRadius: const BorderRadius.all(Radius.circular(12.0))),
+      ),
+    );
+  }
+
+}
